# Optimise solana dockerfiles