--- a/Dockerfile.client
+++ b/Dockerfile.client
@@ -33,6 +33,7 @@ ENV BRIDGE_ADDRESS="Bridge1p5gheXUvJ6jGWGeCsgPKgnE3YgdGKRVCMY9o"
 
 RUN --mount=type=cache,target=/root/.cache \
     --mount=type=cache,target=target \
+    --mount=type=cache,target=/usr/local/cargo/registry,id=cargo_registry \
 	set -xe && \
     cargo build --manifest-path ./Cargo.toml --package bridge_client --release --locked && \
     cargo build --manifest-path ./Cargo.toml --package token_bridge_client --release --locked && \