--- a/solana/.dockerignore
+++ b/solana/.dockerignore
@@ -1,2 +1,7 @@
 bin
 **/target
+**/bundler
+**/nodejs
+artifacts-*
+**/Makefile
+*.md