This PR optimises the solana docker images by caching and removing unnecessary dependencies. Also published prebuilt base images, which make building faster, and more robust against upstream dependency changes.

```
| Dockerfile | before | after |
|------------+--------+-------|
| --no-cache | 320s   | 130s  |
| cache      | 112s   | 25s   |

Code changes will no longer trigger a fully dependency tree rebuild.


<!-- Reviewable:start -->
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/certusone/wormhole/1334)
<!-- Reviewable:end -->