@leoluk 2022-07-14
The alternative way would be to use buildkit caching layers built on dev.v2, so we wouldn't have to hardcode images and rebuild cleanly when desired (such as for releases).

But this works too.